+++ src/realtimeScanners/scanners/containers/containersScannerCommand.ts
@@ -7,7 +7,7 @@ import { ConfigurationManager } from "../../configuration/configurationManager";
 import { constants } from "../../../utils/common/constants";
 import { CxRealtimeEngineStatus } from "@checkmarxdev/ast-cli-javascript-wrapper/dist/main/oss/CxRealtimeEngineStatus";
 import { buildCommandButtons, renderCxAiBadge } from "../../../utils/utils";
-import {ContainersHoverData} from "../../common/types";
+import { ContainersHoverData } from "../../common/types";
 
 export class ContainersScannerCommand extends BaseScannerCommand {
 	constructor(
@@ -57,8 +57,14 @@ export class ContainersScannerCommand extends BaseScannerCommand {
 		const key = `${filePath}:${line}`;
 
 		const hoverData = scanner.getHoverData().get(key);
-		if (!hoverData) {
-			return undefined;
+
+		const diagnostics = scanner.getDiagnosticsMap().get(document.uri.fsPath) ?? [];
+		const hasDiagnostic = diagnostics.some(
+			(d) => d.range.start.line === position.line
+		);
+
+		if (!hoverData || !hasDiagnostic) {
+			return;
 		}
 
 		return this.createHoverContent(hoverData);
@@ -70,7 +76,7 @@ export class ContainersScannerCommand extends BaseScannerCommand {
 		md.isTrusted = true;
 		const { vulnerabilities, ...hoverDataWithoutVuln } = hoverData;
 		const args = encodeURIComponent(JSON.stringify([hoverDataWithoutVuln]));
-		const buttons = buildCommandButtons(args);
+		const buttons = buildCommandButtons(args, true);
 		const isVulnerable = this.isVulnerableStatus(hoverData.status);
 		const isMalicious = hoverData.status === "Malicious";
 
@@ -93,17 +99,17 @@ export class ContainersScannerCommand extends BaseScannerCommand {
 	private renderID(hoverData: ContainersHoverData): string {
 		if (hoverData.status == CxRealtimeEngineStatus.malicious) {
 			return `
-<b>${hoverData.imageName}:${hoverData.imageTag}</b> 
+<b>${hoverData.imageName}:${hoverData.imageTag}</b>
 <i style="color: dimgrey;"> - ${hoverData.status} image <br></i>
 `;
 		}
 		return `
-<b>${hoverData.imageName}:${hoverData.imageTag}</b> 
+<b>${hoverData.imageName}:${hoverData.imageTag}</b>
 <i style="color: dimgrey;"> - ${hoverData.status} severity image <br></i>
 `;
 	}
 
-	private renderImageIcon() : string {
+	private renderImageIcon(): string {
 		return `<img src="https://raw.githubusercontent.com/Checkmarx/ast-vscode-extension/main/media/icons/realtimeEngines/container_image.png" width="15" height="16" style="vertical-align: -12px;"/>`;
 	}
 
@@ -116,8 +122,12 @@ export class ContainersScannerCommand extends BaseScannerCommand {
 		].includes(status as any);
 	}
 
+	private renderMaliciousFinding(): string {
+		return `<img src="https://raw.githubusercontent.com/Checkmarx/ast-vscode-extension/main/media/icons/maliciousFindig.png" style="vertical-align: -12px;" />`;
+	}
+
 	private renderMaliciousIcon(): string {
-		return `<img src="https://raw.githubusercontent.com/Checkmarx/ast-vscode-extension/main/media/icons/malicious.png" width="15" height="16" style="vertical-align: -12px;"/>`;
+		return `<img src="https://raw.githubusercontent.com/Checkmarx/ast-vscode-extension/main/media/icons/malicious.png" width="10" height="11" style="vertical-align: -12px;"/>`;
 	}
 
 	private renderVulnCounts(vulnerabilities: Array<{ severity: string }>): string {